+++ src/main/java/insane96mcp/survivalreimagined/module/experience/feature/EnchantmentsFeature.java
@@ -79,6 +79,9 @@ public class EnchantmentsFeature extends Feature {
 	@Config
 	@Label(name = "Buff Feather Falling", description = "Increases the damage protection from Feather Falling. From 12% per level to 16% per level")
 	public static Boolean buffFeatherFalling = true;
+	@Config
+	@Label(name = "Buff Thorns", description = "Thorns no longer deals damage to items.")
+	public static Boolean buffThorns = true;
 
 	//TODO Make enchantments deactivable
 	public EnchantmentsFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {